Broke up embeds, added tagline to queue embed
