+++ commands/rps/leave.js
@@ -1,6 +1,6 @@
 const { SlashCommandBuilder } = require('discord.js');
 const { findPlayerQueue, removePlayerFromQueue } = require('../../src/game/manageQueues');
-const { leaveEmbed } = require('../../src/utils/embeds');
+const { leave: leaveEmbed } = require('../../src/embeds');
 
 
 module.exports = {
@@ -16,17 +16,17 @@ module.exports = {
             if (!queue) return channel.send({ content: 'You are not in a queue.', ephemeral: true });
 
             if (interaction.inGuild()) {
-                await channel.send({ embeds: [leaveEmbed(queue, user)] });
+                await channel.send({ embeds: [leaveEmbed(user)] });
                 console.log(`${user.username} left Lobby ${queueId}`);
                 return;
             }
 
-            await user.send({ embeds: [leaveEmbed(queue, user)] });
+            await user.send({ embeds: [leaveEmbed(user)] });
             console.log(`${user.username} left Lobby ${queueId}`);
         } else {
             if (!queue) return interaction.reply({ content: 'You are not in a queue.', ephemeral: true });
 
-            await interaction.reply({ embeds: [leaveEmbed(queue, user)] });
+            await interaction.reply({ embeds: [leaveEmbed(user)] });
             console.log(`${user.username} left Lobby ${queueId}`);
         }
     }
