+++ src/embeds/club.js
@@ -1,0 +1,10 @@
+const { EmbedBuilder } = require('discord.js');
+const { footer } = require('./footer');
+
+module.exports = (club) => {
+    const { name, abbreviation } = club;
+    return new EmbedBuilder()
+        .setTitle(name)
+        .setDescription(abbreviation)
+        .setFooter(footer);
+};
