+++ src/embeds/footer.js
@@ -1,0 +1,6 @@
+const footer = {
+    text: "Powered by RPS",
+    iconURL: "https://i.imgur.com/YODnmBn.png"
+};
+
+module.exports = { footer };
