+++ src/embeds/index.js
@@ -1,0 +1,37 @@
+const challenge = require('./challenge');
+const club = require('./club');
+const demotion = require('./demotion');
+const eloResult = require('./eloResult');
+const game = require('./game');
+const gameDraw = require('./gameDraw');
+const gameLoss = require('./gameLoss');
+const gameWin = require('./gameWin');
+const leaderboard = require('./leaderboard');
+const leave = require('./leave');
+const promotion = require('./promotion');
+const queue = require('./queue');
+const ranks = require('./ranks');
+const results = require('./results');
+const stats = require('./stats');
+const status = require('./status');
+const waiting = require('./waiting');
+
+module.exports = {
+    challenge,
+    club,
+    demotion,
+    eloResult,
+    game,
+    gameDraw,
+    gameLoss,
+    gameWin,
+    leaderboard,
+    leave,
+    promotion,
+    queue,
+    ranks,
+    results,
+    stats,
+    status,
+    waiting
+};
