--- src/utils/capitalize.js
@@ -1,3 +1,0 @@
-module.exports = (string) => {
-    return string.charAt(0).toUpperCase() + string.slice(1);
-};
